支付成功

+++ clzmall-app/src/main/java/com/clzmall/app/entity/dto/PayParam.java
@@ -3,6 +3,8 @@ package com.clzmall.app.entity.dto;
 import lombok.Data;
 import lombok.ToString;
 
+import java.math.BigDecimal;
+
 /**
  * Created by jiangjianshi on 18/9/1.
  */
@@ -11,7 +13,7 @@ import lombok.ToString;
 public class PayParam {
 
     private String uid;
-    private double money;
+    private BigDecimal money;
     private String remark;
     private String payName;
     private String nextAction;
